Create a basic gui that assumes cube size and height.
The interface lets the user select a png and save it as an OpenSCAD
project, eliminating the command-line aspect entirely. The interface
includes an image previewer since that felt right to add.

The tool doesn't include cube size and height adjustment, because those
variables can be adjusted in OpenSCAD directly.

I am unlikely to add more to this tool, because in the current form it
is sufficient for my use case.